% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecd_funs.R
\name{oecd_make_filter}
\alias{oecd_make_filter}
\title{Create an OECD Filter String}
\usage{
oecd_make_filter(x)
}
\arguments{
\item{x}{A list of strings representing different filter criteria for the OECD API.}
}
\value{
A single string with the filter criteria concatenated using '+' and '.' as separators.
}
\description{
This function takes a list of filtering arguments and constructs a filter string for an OECD API call.
}
\examples{
make_oecd_filter(list("A", "sna_geo", "", "", "sna6a_transact", "", "sna_activity", "", "", "sna_measures", "", ""))
}
