% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecfin_weights.R
\name{read_ecfin_weights}
\alias{read_ecfin_weights}
\title{Read ecfin weights from zip file}
\usage{
read_ecfin_weights(zipfile, file_pre)
}
\arguments{
\item{zipfile}{A zip-file to read.}

\item{file_pre}{A first part of the filename inside zip-file.}
}
\description{
Weights are in induvidual csv-files.
}
