% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{save_dat}
\alias{save_dat}
\title{Save a dataset to inst/extdata as Parquet (for package build time)}
\usage{
save_dat(
  x,
  filename = deparse(substitute(x)),
  dir = "inst/extdata",
  overwrite = FALSE
)
}
\arguments{
\item{x}{A \code{data.frame} (or tibble) to be saved.}

\item{filename}{File name to write. Defaults to the name of \code{x} with a
\code{.parquet} extension.}

\item{dir}{Directory where to write during development. Defaults to
\code{"inst/extdata"}. The directory is created if it does not exist.}

\item{overwrite}{Logical, overwrite an existing file (default \code{FALSE}).}
}
\value{
(Invisibly) the path to the written file.
}
\description{
Writes a data frame to the package's \code{inst/extdata/} directory so that,
after installation, the file will be available under \code{system.file("extdata", ...)}.
Intended to be used during development (not at runtime after installation).
}
\examples{
\dontrun{
# During development of the package:
save_dat(mtcars) # writes inst/extdata/mtcars.parquet
}
}
